+++ shared-common-graph/src/commonMain/kotlin/learn/git/common/graph/walker/WalkerConfig.kt
@@ -1,8 +1,8 @@
 package learn.git.common.graph.walker
 
 data class WalkerConfig(
-    val siblingSeparation: Int,
-    val levelSeparation: Int,
-    val subtreeSeparation: Int,
-    val nodeSize: Int
+    val siblingSeparation: Float,
+    val levelSeparation: Float,
+    val subtreeSeparation: Float,
+    val nodeSize: Float
 )
